+++ event_test.go
@@ -2,7 +2,6 @@ package budget
 
 import (
 	"errors"
-	"github.com/dottics/dutil"
 	"github.com/google/uuid"
 	"github.com/johannesscr/micro/microtest"
 	"testing"
@@ -10,21 +9,18 @@ import (
 )
 
 func TestService_CreateEvent(t *testing.T) {
-	type E struct {
-		event Event
-		e     dutil.Error
-	}
 	tt := []struct {
 		name     string
 		UUID     uuid.UUID
-		event    EventCreate
+		payload  EventCreate
 		exchange *microtest.Exchange
-		E        E
+		event    Event
+		e        error
 	}{
 		{
 			name: "403 Permission Required",
 			UUID: uuid.MustParse("ae9f5130-81fe-4526-9573-f7e892cc2e01"),
-			event: EventCreate{
+			payload: EventCreate{
 				Name:      "test event one",
 				Amount:    12.19,
 				Debit:     true,
@@ -35,62 +31,16 @@ func TestService_CreateEvent(t *testing.T) {
 			exchange: &microtest.Exchange{
 				Response: microtest.Response{
 					Status: 403,
-					Body: `{
-						"message":"Forbidden: unable to process request",
-						"data":{},
-						"errors":{
-							"permission":["Please ensure you have permission"]
-						}
-					}`,
+					Body:   noPermission,
 				},
 			},
-			E: E{
-				event: Event{},
-				e: &dutil.Err{
-					Status: 403,
-					Errors: map[string][]string{
-						"permission": {"Please ensure you have permission"},
-					},
-				},
-			},
-		},
-		{
-			name: "500 Internal Server Error",
-			UUID: uuid.MustParse("ae9f5130-81fe-4526-9573-f7e892cc2e01"),
-			event: EventCreate{
-				Name:      "test event two",
-				Amount:    12.19,
-				Debit:     true,
-				Credit:    false,
-				StartDate: time.Date(2021, 11, 18, 0, 0, 0, 0, time.UTC),
-				EndDate:   time.Date(2021, 11, 19, 0, 0, 0, 0, time.UTC),
-			},
-			exchange: &microtest.Exchange{
-				Response: microtest.Response{
-					Status: 500,
-					Body: `{
-						"message":"InternalServerError: unable to process request",
-						"data":{},
-						"errors":{
-							"internal_server_error":["some random internal error"]
-						}
-					}`,
-				},
-			},
-			E: E{
-				event: Event{},
-				e: &dutil.Err{
-					Status: 500,
-					Errors: map[string][]string{
-						"internal_server_error": {"some random internal error"},
-					},
-				},
-			},
+			event: Event{},
+			e:     errors.New("no permission"),
 		},
 		{
 			name: "200 Successful",
 			UUID: uuid.MustParse("ae9f5130-81fe-4526-9573-f7e892cc2e01"),
-			event: EventCreate{
+			payload: EventCreate{
 				Name:      "test event three",
 				Amount:    12.19,
 				Debit:     true,
@@ -100,7 +50,7 @@ func TestService_CreateEvent(t *testing.T) {
 			},
 			exchange: &microtest.Exchange{
 				Response: microtest.Response{
-					Status: 200,
+					Status: 201,
 					Body: `{
 						"message":"event created successfully",
 						"data":{
@@ -119,18 +69,16 @@ func TestService_CreateEvent(t *testing.T) {
 					}`,
 				},
 			},
-			E: E{
-				event: Event{
-					UUID:      uuid.MustParse("b86768ee-69de-4fb2-81eb-ab96d14e37ae"),
-					Name:      "test event three",
-					Amount:    12.19,
-					Debit:     true,
-					Credit:    false,
-					StartDate: time.Date(2021, 11, 18, 0, 0, 0, 0, time.UTC),
-					EndDate:   time.Date(2021, 11, 19, 0, 0, 0, 0, time.UTC),
-				},
-				e: nil,
+			event: Event{
+				UUID:      uuid.MustParse("b86768ee-69de-4fb2-81eb-ab96d14e37ae"),
+				Name:      "test event three",
+				Amount:    12.19,
+				Debit:     true,
+				Credit:    false,
+				StartDate: time.Date(2021, 11, 18, 0, 0, 0, 0, time.UTC),
+				EndDate:   time.Date(2021, 11, 19, 0, 0, 0, 0, time.UTC),
 			},
+			e: nil,
 		},
 	}
 
@@ -143,20 +91,16 @@ func TestService_CreateEvent(t *testing.T) {
 			// add budget-micro-service exchange
 			ms.Append(tc.exchange)
 
-			ev, e := s.CreateEvent(tc.event)
+			ev, e := s.CreateEvent(tc.payload)
 
 			// test errors
-			if tc.E.e != nil {
-				if tc.E.e.Error() != e.Error() {
-					t.Errorf("expected error '%s' got '%s'", tc.E.e.Error(), e.Error())
-				}
-			} else if e != nil {
-				t.Errorf("unexpected error: %s", e.Error())
+			if NotEqualError(e, tc.e) {
+				t.Errorf("expected error '%v' got '%v'", tc.e, e)
 			}
 
 			// test event
-			if ev != tc.E.event {
-				t.Errorf("expected an event '%v' got '%v'", tc.E.event, ev)
+			if ev != tc.event {
+				t.Errorf("expected an event\n'%+v'\ngot\n'%+v'", tc.event, ev)
 			}
 
 			// TODO: test exchange request body
@@ -165,20 +109,19 @@ func TestService_CreateEvent(t *testing.T) {
 }
 
 func TestUpdateEvent(t *testing.T) {
-	type E struct {
-		event Event
-		e     error
-	}
 	tt := []struct {
 		name     string
 		uuid     uuid.UUID
-		event    EventUpdate
+		payload  EventUpdate
 		exchange *microtest.Exchange
-		E        E
+		uri      string
+		event    Event
+		e        error
 	}{
 		{
 			name: "error",
-			event: EventUpdate{
+			uuid: uuid.MustParse("8feec066-2dfb-44f5-b353-1cb6e75c3084"),
+			payload: EventUpdate{
 				ItemUUID:  uuid.MustParse("b86768ee-69de-4fb2-81eb-ab96d14e37ae"),
 				Name:      "test event",
 				Debit:     true,
@@ -192,14 +135,14 @@ func TestUpdateEvent(t *testing.T) {
 					Body:   noPermission,
 				},
 			},
-			E: E{
-				event: Event{},
-				e:     errors.New("no permission"),
-			},
+			uri:   "/event/8feec066-2dfb-44f5-b353-1cb6e75c3084",
+			event: Event{},
+			e:     errors.New("no permission"),
 		},
 		{
 			name: "success",
-			event: EventUpdate{
+			uuid: uuid.MustParse("0aab721b-4224-474c-9df8-e77117a31a02"),
+			payload: EventUpdate{
 				ItemUUID:  uuid.MustParse("b86768ee-69de-4fb2-81eb-ab96d14e37ae"),
 				Name:      "test event",
 				Debit:     false,
@@ -215,7 +158,7 @@ func TestUpdateEvent(t *testing.T) {
 						"message":"event updated successfully",
 						"data":{
 							"event":{
-								"uuid":"b86768ee-69de-4fb2-81eb-ab96d14e37ae",
+								"uuid":"0aab721b-4224-474c-9df8-e77117a31a02",
 								"name":"test event",
 								"debit":false,
 								"credit":true,
@@ -229,18 +172,17 @@ func TestUpdateEvent(t *testing.T) {
 					}`,
 				},
 			},
-			E: E{
-				event: Event{
-					UUID:      uuid.MustParse("b86768ee-69de-4fb2-81eb-ab96d14e37ae"),
-					Name:      "test event",
-					Debit:     false,
-					Credit:    true,
-					Amount:    125.67,
-					StartDate: time.Date(2021, 11, 18, 0, 0, 0, 0, time.UTC),
-					EndDate:   time.Date(2021, 11, 19, 0, 0, 0, 0, time.UTC),
-				},
-				e: nil,
+			uri: "/event/0aab721b-4224-474c-9df8-e77117a31a02",
+			event: Event{
+				UUID:      uuid.MustParse("0aab721b-4224-474c-9df8-e77117a31a02"),
+				Name:      "test event",
+				Debit:     false,
+				Credit:    true,
+				Amount:    125.67,
+				StartDate: time.Date(2021, 11, 18, 0, 0, 0, 0, time.UTC),
+				EndDate:   time.Date(2021, 11, 19, 0, 0, 0, 0, time.UTC),
 			},
+			e: nil,
 		},
 	}
 
@@ -253,33 +195,30 @@ func TestUpdateEvent(t *testing.T) {
 			// add budget-micro-service exchange
 			ms.Append(tc.exchange)
 
-			ev, e := s.UpdateEvent(tc.uuid, tc.event)
+			ev, e := s.UpdateEvent(tc.uuid, tc.payload)
 			// test error
-			if tc.E.e != nil {
-				if tc.E.e.Error() != e.Error() {
-					t.Errorf("expected error '%s' got '%s'", tc.E.e.Error(), e.Error())
-				}
-			} else if e != nil {
-				t.Errorf("unexpected error: %s", e.Error())
+			if NotEqualError(e, tc.e) {
+				t.Errorf("expected error '%v' got '%v'", tc.e, e)
 			}
 			// test event
-			if tc.E.event != ev {
-				t.Errorf("expected event '%+v' got '%+v'", tc.E.event, ev)
+			if tc.event != ev {
+				t.Errorf("expected event '%+v' got '%+v'", tc.event, ev)
+			}
+			// test exchange request uri
+			if tc.uri != tc.exchange.Request.RequestURI {
+				t.Errorf("expected uri '%s' got '%s'", tc.uri, tc.exchange.Request.RequestURI)
 			}
 		})
 	}
 }
 
 func TestService_DeleteEvent(t *testing.T) {
-	type E struct {
-		e        error
-		exReqURI string
-	}
 	tt := []struct {
 		name     string
 		UUID     uuid.UUID
 		exchange *microtest.Exchange
-		E        E
+		uri      string
+		e        error
 	}{
 		{
 			name: "403 Bad Request",
@@ -290,10 +229,8 @@ func TestService_DeleteEvent(t *testing.T) {
 					Body:   noPermission,
 				},
 			},
-			E: E{
-				exReqURI: "/event/7dee09f0-2ba2-4b10-9c88-0c973ad4ebd0",
-				e:        errors.New("no permission"),
-			},
+			uri: "/event/7dee09f0-2ba2-4b10-9c88-0c973ad4ebd0",
+			e:   errors.New("no permission"),
 		},
 		{
 			name: "200 Success",
@@ -307,10 +244,8 @@ func TestService_DeleteEvent(t *testing.T) {
 					}`,
 				},
 			},
-			E: E{
-				exReqURI: "/event/2c7b7d76-c9e0-49f8-b585-166ac70dba6f",
-				e:        nil,
-			},
+			uri: "/event/2c7b7d76-c9e0-49f8-b585-166ac70dba6f",
+			e:   nil,
 		},
 	}
 
@@ -324,16 +259,12 @@ func TestService_DeleteEvent(t *testing.T) {
 			ms.Append(tc.exchange)
 
 			e := s.DeleteEvent(tc.UUID)
-			if tc.E.e != nil {
-				if tc.E.e.Error() != e.Error() {
-					t.Errorf("expected error '%v' got '%v'", tc.E.e.Error(), e.Error())
-				}
-			} else if e != nil {
-				t.Errorf("unpexpected error: %s", e.Error())
+			if NotEqualError(e, tc.e) {
+				t.Errorf("expected error '%v' got '%v'", tc.e, e)
 			}
-			URI := tc.exchange.Request.RequestURI
-			if URI != tc.E.exReqURI {
-				t.Errorf("expected URI '%s' got '%s'", tc.E.exReqURI, URI)
+
+			if tc.exchange.Request.RequestURI != tc.uri {
+				t.Errorf("expected URI '%s' got '%s'", tc.uri, tc.exchange.Request.RequestURI)
 			}
 		})
 	}